Adjust enum representation and in doing so deal with circular imports annoying
