+++ js/src/ast/patterns/ListPattern.ts
@@ -1,7 +1,8 @@
 import { CompilationScope } from '../../compiler/CompilationScope'
 import { ErrorType } from '../../type-checker/errors/Error'
 import { list } from '../../type-checker/types/builtins'
-import { AliasSpec, NType, unknown } from '../../type-checker/types/types'
+import { NType, unknown } from '../../type-checker/types/types'
+import { AliasSpec } from '../../type-checker/types/TypeSpec'
 import schema, * as schem from '../../utils/schema'
 import { Base, BasePosition } from '../base'
 import {
